--- src/integrators/bdpt/mod.rs
@@ -1,7 +1,0 @@
-pub mod integrator;
-pub mod scoped_assignment;
-pub mod subpath;
-pub mod vertex;
-pub mod vertex_interaction;
-
-pub use integrator::*;
