+++ src/integrators/directlighting.rs
