--- src/integrators/mlt/mod.rs
@@ -1,4 +1,0 @@
-pub mod integrator;
-pub mod sampler;
-
-pub use integrator::*;
