+++ src/integrators/mod.rs
@@ -6,16 +6,17 @@ pub mod directlighting;
 pub mod mlt;
 pub mod path;
 pub mod sppm;
+pub mod subpath;
 pub mod volpath;
 pub mod whitted;
 
-pub use ao::integrator::*;
-pub use aov::integrator::*;
-pub use bdpt::integrator::*;
+pub use ao::*;
+pub use aov::*;
+pub use bdpt::*;
 pub use create_integrator::*;
-pub use directlighting::integrator::*;
-pub use mlt::integrator::*;
-pub use path::integraor::*;
-pub use sppm::integrator::*;
-pub use volpath::integrator::*;
-pub use whitted::integrator::*;
+pub use directlighting::*;
+pub use mlt::*;
+pub use path::*;
+pub use sppm::*;
+pub use volpath::*;
+pub use whitted::*;
