--- src/integrators/path/mod.rs
@@ -1,3 +1,0 @@
-pub mod integraor;
-
-pub use integraor::*;
