--- src/integrators/sppm/mod.rs
@@ -1,3 +1,0 @@
-pub mod integrator;
-
-pub use integrator::*;
