+++ src/integrators/subpath/generate_subpath.rs
