+++ src/integrators/subpath/scoped_assignment.rs
