+++ src/integrators/subpath/vertex_interaction.rs
