+++ src/integrators/whitted.rs
